monitor/requirements.txt
quilt==2.8.0
gitdb2==2.0.3
